+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Notable changes to this project.
 ## dev
 - adding `fncPercentile`, `mmcPercentile` & `corrPercentile` to `daily_submissions_performances`
 - replace deprecated GraphQL `v2UserProfile` call with `v3UserProfile`
+- new `daily_model_performances`, replacing `daily_user_performances`
 
 ## [2.6.0] - 2021-07-12
 - cli: remove deprecated 'payments' and 'user-activities' commands (#51)