+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Notable changes to this project.
 - replace deprecated GraphQL `signalsUserProfile` call with `v2SignalsProfile`
 - new `daily_model_performances`, replacing `daily_user_performances`
 - signals: new `daily_model_performances`, replacing `daily_user_performances`
+- update & fix command line interface
 
 ## [2.6.0] - 2021-07-12
 - cli: remove deprecated 'payments' and 'user-activities' commands (#51)